rtc: Disable alarm irq if alarm time is in the past
If device is boot up by rtc alarm, the alarm irq will still be
enabled and the alarm time is smaller than current rtc time before
any alarm is set or canceled. If device is powered off now, it will
boot up automatically as the alarm irq is enabled.

So disable alarm irq if alarm is enabled and alarm time is in the
past.

Change-Id: I53dc7cdb1fa7c56eea07ed01d1f6cb63f0bb8a8d
Signed-off-by: Guixiong Wei <guixiong@codeaurora.org>